backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ArthurGolubev <gg.arthur.go@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
neo4j = "^5.8.0"
fastapi = {extras = ["all"], version = "^0.95.1"}
loguru = "^0.7.0"
python-cypher = "^0.20.1"
sqlmodel = "^0.0.8"
alembic = "^1.10.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyhumps = "^3.8.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
